+++ docs/source/changelog.md
@@ -1,6 +1,8 @@
 # Changelog
 
-## 0.12.12 (2026-02-xx)
+## 0.12.12 (2026-02-XX)
+- Fix dataframe upload fallbacks for MSSQL and Snowflake backends
+- Fix crash when prod schema could not be moved to transaction schema (e.g. due to broken views)
 - Fix: improved google cloud storage (GCS) support
 
 ## 0.12.11 (2026-01-23)